[VR] Initial support for VR via OpenXR
This commit adds initial support for playing Quake 3 Arena in VR using
OpenXR API. This work was initially based on Quake3Quest project which
was rebased on top of newest `ioquake3` idTech3 fork, adjusted for PCVR
(including rewriting big chunks of OpenXR-related code) with a lot of
bugs fixed and many features and QoL improvements added along the way.

Currently building and playing is supported only on Windows (with MSVC
or Mingw).