[#1][#7] Fix delay/wobbly effect when quickly moving head
This is UX improvement/optimization. Previously the game loop worked by
updating the game state and then rendering frame, which updted the
XR-related state such as HMD and controllers poses. This unfortunately
led to the problem that frame state update were using XR data from
previous frames which caused 1-frame delay which was a problem reported
by multiple players.

This logic reworks the game loop to wrap whole frame processing in XR
BeginFrame and EndFrame so that state update also uses new logic.
Unfortunately this complicates VR Renderer's logic a bit due to the fact
that `Com_Frame()` can longjmp() after reinitializing the whole context
(GL, window, etc.) which means we need to reinitialize OpenXR as well.
So it may be possible, that after calling `xrWaitFrame`, `xrBeginFrame`
and `Com_Frame`, we will end up (for the rest of the `Com_Frame`)
without acquired frame. To fix this, reinitialization logic now can redo
the logic done prior to calling `Com_Frame` through the
`VR_Renderer_RestoreState` function (this time on new XR instance).

It seems that this new game loop logic also fixes the problem described
in #7 and reinitialization now is much smoother and doesn't cause
stuttering in either the game or SteamVR/VR runtime itself anymore.