Make vr.virtual_screen more intuitive (?)
Of two minds about this commit. Basically, wanted to make it so that
vr.virtual_screen is true any time we are rendering the virtual screen
(including in first-person follow mode), but this of course means adding
vr.first_person_following, because sometimes behavior differs between
these cases.

The thing is, I felt like vr.virtual_screen was a trap, since it didn't
_always_ indicate we would be rendering in a virtual screen.