Allow toggling desktop mirror mode
Can toggle from off / windowed / fullscreen now. In windowed mode, will
vid_restart when changing from one to both eyes, to avoid unnecessary
pillarboxing.

Intentional change, but should find out if Ripper likes this or not: the
old mirroring mode aimed to horizontally fill the scree, but this cut
off vertically, making things like the floating HUD scoreboard and the
console displays not show up properly. If we need to make this a toggle,
we can.