Some menu-related fixes/improvements
- Make the in-menu view freeze the camera position/angle
- Fix incorrect color selector on team arena in-game menu

Unrelated, but also in this commit is a fix for single-player
VRFM_THIRDPERSON_2 follow that was causing slight movement in the camera
even when the player isn't moving their head.

The end result: a hopefully less disorienting effect when in a "virtual
theater" that is displaying a screen, so that the screen moves around,
but its contents don't. Well, in a public game the other players keep
doing their thing, but your view will stay consistent.

Works across all follow modes, deathcam, spectator, etc.