[#25] Added in-game option for supersampling (in [0.5, 2.0] range)
This commit adds new configuration option for enabling supersampling
(or subsampling) to improve visual quality (much crispier image) or
performance. Available options are in [0.5, 2.0] range.

The initial idea was to handle user switching this via SteamVR runtime
config but that wouldn't work for non-SteamVR users. This generic option
should work for everyone.