[#16] Add a notification when newer version is available
This commit adds a feature that checks if latest release on GitHub is
newer than the current build and if so, a new notification menu/screen
is shown to the user with (optionally) option to open the game's page to
make it easier to download newer version for people.

The check itself is based on an API request made to GitHub to find out
information about the latest release on GitHub (RippeR37/q3vr) and
its version is parsed from `tag_name` field. The expected format is:

  [v]major[.minor[.patch]]

e.g. `v1.2.14` or `2.6`.

On the game side, the UI VM triggers this check, a new logic in expanded
networking stack is implemented to allow in-memory data fetching
(instead of storing it to FS) and when the request finishes, we update
cvars which are picked up when next event in main menu is processed.

The logic is limited to execute only once per game start.