[#27] Fix Virtual Screen rotating when not in a follow mode
This commit adjusts behavior of Virtual Screen when in `fixed` mode
(instead of `follow`). Previously VS was always rotating exactly towards
the position of player's head/HMD, even if it was in fixed mode, which
was somewhat confusing to players. With this change, it will only rotate
when in following mode or when player manually rests its position and
orientation via a button press.