Fix rendering text shadow in HUD
Recently we've increased rendering resolution of HUD from 640x480 to
1280x960 (doubled). Unfortunately the logic that computes where to
render text shadow on the screen (/HUD) was hardcoded for 640x480 which
resulted in bigger text having its shadow computed way too low/to the
right. With this fix all text shadows should now be rendered correctly
which should improve readability of the text on screen.